moved view to participant side
